--- components/Accent/styles.module.scss
@@ -1,8 +1,0 @@
-@import '../../variables.scss';
-
-.root {
-  height: 18px;
-  width: 16px;
-  border-bottom: solid $black 5px;
-  border-left: solid $black 5px;
-}
