--- components/Button/styles.module.scss
@@ -1,28 +1,0 @@
-@import '../../variables.scss';
-
-.button {
-  padding: 16px 30px;
-  text-align: center;
-  font-size: 18px;
-  font-weight: bold;
-  border-radius: $border-radius;
-  text-decoration: none;
-  letter-spacing: 1px;
-  &:hover {
-    text-decoration: none;
-  }
-  &.primary {
-    color: $white;
-    background-color: $primary-dark;
-    &:hover {
-      background-color: lighten($color: $primary-dark, $amount: 5);
-    }
-  }
-  &.secondary {
-    color: $primary-dark;
-    background-color: $secondary;
-    &:hover {
-      background-color: darken($color: $secondary, $amount: 20);
-    }
-  }
-}
