--- components/Header/styles.module.scss
@@ -1,26 +1,0 @@
-@import '../../variables.scss';
-
-.content {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  @include container;
-  @include screen('tablet') {
-    flex-direction: row;
-    align-items: center;
-  }
-}
-
-.divider {
-  width: 100%;
-  height: 16px;
-  background: linear-gradient(
-    110deg,
-    rgba(179, 241, 253, 1) 0%,
-    rgba(166, 237, 252, 1) 11%,
-    rgba(209, 251, 255, 1) 11%,
-    rgba(181, 242, 253, 1) 57%,
-    rgba(194, 255, 233, 1) 57%,
-    rgba(186, 255, 233, 1) 100%
-  );
-}
