--- components/Nav/styles.module.scss
@@ -1,45 +1,0 @@
-@import '../../variables.scss';
-
-.navContainer {
-  @include container;
-  font-size: 20px;
-
-  ul {
-    list-style: none;
-    display: flex;
-    flex-direction: row;
-    align-items: center;
-    padding: 0;
-
-    li + li {
-      margin-left: 40px;
-    }
-
-    li:first-of-type {
-      margin-right: auto;
-    }
-
-    li {
-      a,
-      a:visited {
-        color: $black;
-        text-decoration: none;
-        padding-top: 20px;
-        padding-bottom: 20px;
-      }
-      &.activeRoute {
-        a {
-          font-weight: bold;
-          color: blue;
-        }
-      }
-    }
-  }
-}
-
-.image {
-  //placeholder
-  width: 50px;
-  height: 50px;
-  background: blue;
-}
