+++ components/scoped/apply/FAQCell/index.js
