+++ components/scoped/apply/FAQCell/styles.module.scss
@@ -1,0 +1,32 @@
+@import '../../../../variables.scss';
+
+.root {
+  @include screen('phone') {
+    display: flex;
+  }
+}
+
+.icon {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  padding: 5px 15px;
+  color: $black;
+  background-color: $secondary;
+  height: 50px;
+  width: 50px;
+  margin-right: 30px;
+  border-radius: 50%;
+  font-size: 20px;
+  margin-bottom: 20px;
+
+  @include screen('phone') {
+    margin-bottom: 0;
+  }
+}
+
+.question {
+  display: flex;
+  font-size: 20px;
+  font-weight: bold;
+}
