+++ components/scoped/work/ChapterTable/styles.module.scss
@@ -1,0 +1,45 @@
+@import '../../../../variables.scss';
+
+.root {
+  display: flex;
+  flex-direction: column;
+}
+
+.chapters {
+  overflow-x: scroll;
+  white-space: nowrap;
+  width: 100%;
+  overflow-x: scroll;
+  border-bottom: solid 1px lightgray;
+  margin-bottom: 12px;
+  .chapter_select:first-child {
+    margin-left: 0;
+  }
+}
+
+.chapter_select {
+  margin: 12px 12px 0 12px;
+  display: inline-block;
+  *display: inline; /* For IE7*/
+  *zoom: 1; /* For IE7*/
+  vertical-align: top;
+  cursor: pointer;
+  &.selected {
+    border-bottom: solid 5px $accent;
+  }
+  &:focus {
+    outline: 0;
+  }
+}
+
+.logo {
+  width: 65px;
+  height: 65px;
+  object-fit: contain;
+}
+
+.highlight {
+  background-color: $accent;
+  width: 65px;
+  height: 10px;
+}
