+++ components/shared/Accent/index.js
