+++ components/shared/Container/styles.module.scss
@@ -1,0 +1,11 @@
+@import '../../../variables.scss';
+
+  .root {
+   margin-bottom: 50px;
+   @include screen('tablet') {
+     margin-bottom: 100px;
+   }
+   @include screen('desktop-wide') {
+     margin-bottom: 150px;
+   }
+ }
