+++ components/shared/Footer/styles.module.scss
@@ -1,0 +1,92 @@
+@import '../../../variables.scss';
+
+.footer_wrapper {
+  background-color: $background;
+}
+
+.footer_content {
+  @include container;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  padding-bottom: 6vw;
+  padding-top: 2vw;
+
+  h2 {
+    margin-bottom: 16px;
+  }
+
+  h4 {
+    display: block;
+    margin-bottom: 16px;
+  }
+
+  p {
+    display: block;
+    margin-top: 0px;
+    margin-bottom: 12px;
+    color: #2e5987;
+  }
+
+  a {
+    display: block;
+  }
+}
+
+.left_content {
+  display: flex;
+  flex-direction: column;
+}
+
+.right_content {
+  display: flex;
+  flex-direction: row;
+
+  a {
+    color: #2e5987;
+  }
+}
+
+.footer_email {
+  color: $accent;
+  font-size: $h4;
+  margin-top: 16px;
+}
+
+.footer_link {
+  color: $black;
+  padding-bottom: 12px;
+}
+
+.link_list {
+  margin-left: 128px;
+}
+
+@media only screen and (max-width: 980px) {
+  .link_list {
+    margin-left: 64px;
+  }
+}
+
+@media only screen and (max-width: 780px) {
+  .right_content {
+    display: flex;
+    flex-direction: column;
+  }
+}
+
+@media only screen and (max-width: 582px) {
+  .link_list {
+    margin-left: 0px;
+  }
+
+  .left_content {
+    padding-bottom: 32px;
+  }
+
+  .footer_content {
+    display: flex;
+    flex-direction: column;
+    padding-bottom: 48px;
+  }
+}
