+++ components/shared/ImageHighlight/styles.module.scss
@@ -1,0 +1,24 @@
+@import '../../../variables.scss';
+
+.root {
+  @include dotted-shadow;
+}
+
+.accent_wrapper {
+  position: absolute;
+  top: -10px;
+  left: -10px;
+}
+
+.accent_1 {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+}
+
+.accent_2 {
+  position: absolute;
+  top: 0;
+  right: 0;
+  transform: rotate(180deg) !important;
+}
