+++ components/shared/Team/dialog-content.js
