+++ pages/_app.js
@@ -1,8 +1,8 @@
 import '@reach/dialog/styles.css';
 import '../global.scss';
 import PropTypes from 'prop-types';
-import Nav from '../components/Nav';
-import Footer from '../components/Footer';
+import Nav from '../components/shared/Nav';
+import Footer from '../components/shared/Footer';
 
 // This default export is required in a new `pages/_app.js` file.
 export default function MyApp({ Component, pageProps }) {
