+++ pages/about/index.js
@@ -1,4 +1,4 @@
-import Header from '../../components/Header';
+import Header from '../../components/shared/Header';
 
 function About() {
   return (
